refactored toolbar and gate
