+++ client/src/gate.tsx
@@ -1,0 +1,29 @@
+import React, { useState, ReactNode, useEffect } from 'react';
+import './App.css';
+import './circuitboard.css';
+import './toolbar.css';
+import Toolbar from './toolbar';
+import {DndContext} from '@dnd-kit/core';
+import {useDraggable, useDroppable} from '@dnd-kit/core';
+import {CSS} from '@dnd-kit/utilities';
+import axios from 'axios';
+
+function Gate(props:any) {
+    const {attributes, listeners, setNodeRef, transform} = useDraggable({
+        id: props.id,
+      });
+      const style = {
+        transform: CSS.Translate.toString(transform),
+        width: 50,
+        height: 50
+        
+      };
+      
+      return (
+        <button ref={setNodeRef} style={style} {...listeners} {...attributes}>
+          <h1>{props.name}</h1>
+        </button>
+      );
+  }
+
+export default Gate;
